Only string match of requestBody
Initially we thought of keeping requestBody as flattened object and requestBodyStr as the string version and execute two queries simultaneously.

While that seemed to be a more robust approach, the implementation would take some time as we would have to detect if the incoming request body is json parsable or not and then populate the flattened object based on that.

For the time being we are only matching against full string